feat(settings): extract the four Git pages
Why: Commit, Remote & sync and Integrations move behind the registry.

Pull & fetch splits into three cards — pull, push and rebase — because
'Move dependent branches' is a rebase setting that had been filed under
pull, and one card per concern is what makes the page readable.

Integrations declares itself dynamic: the forge host list is data, so
there are no fixed rows to index.
