+++ src/screens/Settings.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import {
   PGButton,
-  PGButtonGroup,
   PGIcon,
   PGInput,
-  PGSelect,
   PGToggle,
   pgConfirm,
   pgFlash,
@@ -14,29 +12,22 @@ import {
   type SettingsImportReport,
 } from "@/features/settings/useSettingsStore";
 import {
-  DEFAULT_TICKET_PATTERN,
-  isValidTicketPattern,
-} from "@/features/commits/message";
-import type { UpdateRefsMode } from "@/lib/types";
-import { IdentityForm } from "@/features/commits/identity/IdentityForm";
-import { SavedIdentities } from "@/features/commits/identity/SavedIdentities";
-import { useRepoStore } from "@/features/repo/useRepoStore";
-import {
   SettingsCard,
   SettingsRow,
 } from "@/features/settings/layout/SettingsCard";
 import { AppearancePage } from "@/features/settings/pages/appearance";
+import { CommitPage } from "@/features/settings/pages/commit";
 import { DiffPage } from "@/features/settings/pages/diff";
+import { IntegrationsPage } from "@/features/settings/pages/integrations";
 import { KeyboardPage } from "@/features/settings/pages/keyboard";
+import { RemotePage } from "@/features/settings/pages/remote";
 import { UpdatesPage } from "@/features/settings/pages/updates";
-import { ForgeSettings } from "@/features/forge/ForgeSettings";
 import {
   cliShimStatus,
   diagnosticsReport,
   installCliShim,
   readLogTail,
   revealLogFile,
-  type PullMode,
 } from "@/lib/tauri";
 import { appErrorMessage } from "@/lib/errors";
 import type {
@@ -100,50 +91,22 @@ export function SettingsScreen() {
           Preferences are saved locally and apply to every repository.
         </p>
 
-        <IdentitySection />
+        <CommitPage />
 
         <AppearancePage />
 
+        {/*
+         * TASK 6 TRANSIENT STATE: this card is what remains of "Pull & fetch"
+         * after Task 5 moved every other row into RemotePage. `workspace.watch`
+         * and `workspace.shell` stay here — they belong on the Workspace page
+         * Task 6 adds, not on Remote & sync — so this card is left with just its
+         * header and these two rows until Task 6 empties it for good.
+         */}
         <SettingsCard
           id="pull"
           title="Pull & fetch"
           subtitle="How platypusgit updates your local branches from their upstream."
         >
-          <SettingsRow
-            id="pull.mode"
-            label="Default pull mode"
-            hint={
-              <>
-                <strong>Rebase</strong> replays your local commits on top of
-                origin (linear history).{" "}
-                <strong>Merge</strong> creates a merge commit.{" "}
-                <strong>Fast-forward only</strong> refuses to pull if your branch has diverged.
-              </>
-            }
-            control={
-              <PGButtonGroup
-                size="sm"
-                value={s.defaultPullMode}
-                onChange={(v) => s.set("defaultPullMode", v as PullMode)}
-                options={[
-                  { value: "Rebase", label: "Rebase" },
-                  { value: "Merge", label: "Merge" },
-                  { value: "FastForward", label: "FF-only" },
-                ]}
-              />
-            }
-          />
-          <SettingsRow
-            id="pull.autostash"
-            label="Auto-stash before pull"
-            hint="Stash dirty changes, pull, then pop the stash. Prevents the 'uncommitted changes' error."
-            control={
-              <PGToggle
-                checked={s.autoStashBeforePull}
-                onChange={(v) => s.set("autoStashBeforePull", v)}
-              />
-            }
-          />
           <SettingsRow
             id="workspace.watch"
             label="Watch the working copy"
@@ -170,153 +133,13 @@ export function SettingsScreen() {
               />
             }
           />
-          <SettingsRow
-            id="fetch.auto"
-            label="Auto-fetch"
-            hint="Periodically run fetch in the background so ahead/behind counts stay fresh."
-            control={
-              <PGToggle
-                checked={s.autoFetchEnabled}
-                onChange={(v) => s.set("autoFetchEnabled", v)}
-              />
-            }
-          />
-          <SettingsRow
-            id="fetch.interval"
-            label="Auto-fetch interval"
-            hint="Minutes between background fetches."
-            control={
-              <PGInput
-                type="number"
-                value={String(s.autoFetchMinutes)}
-                onChange={(v) => {
-                  const n = Math.max(1, Math.min(60, parseInt(v, 10) || 5));
-                  s.set("autoFetchMinutes", n);
-                }}
-                style={{ width: 72 }}
-                disabled={!s.autoFetchEnabled}
-              />
-            }
-          />
-          <SettingsRow
-            id="rebase.updateRefs"
-            label="Move dependent branches"
-            hint="When rebasing, also move branches whose tips sit inside the range being replayed — git's rebase --update-refs, the thing that keeps a stack of small PRs from being orphaned. Follow git config uses this repository's own rebase.updateRefs. You are always asked first, and told which branches will move."
-            control={
-              <PGSelect
-                data-testid="rebase-update-refs"
-                value={s.rebaseUpdateRefs}
-                onChange={(v) => s.set("rebaseUpdateRefs", v as UpdateRefsMode)}
-                options={[
-                  { value: "config", label: "Follow git config" },
-                  { value: "always", label: "Always" },
-                  { value: "never", label: "Never" },
-                ]}
-              />
-            }
-          />
-          <SettingsRow
-            id="fetch.prune"
-            label="Prune on fetch"
-            hint="Remove local refs whose upstream branches have been deleted on the remote."
-            control={
-              <PGToggle
-                checked={s.pruneOnFetch}
-                onChange={(v) => s.set("pruneOnFetch", v)}
-              />
-            }
-          />
         </SettingsCard>
 
-        <SettingsCard
-          id="push"
-          title="Push safety"
-          subtitle="Guardrails around destructive remote operations."
-        >
-          <SettingsRow
-            id="push.confirmForce"
-            label="Confirm force-push"
-            hint="Ask for confirmation before a force or force-with-lease push."
-            control={
-              <PGToggle
-                checked={s.confirmForcePush}
-                onChange={(v) => s.set("confirmForcePush", v)}
-              />
-            }
-          />
-        </SettingsCard>
-
-        <SettingsCard
-          id="commit"
-          title="Commit"
-          subtitle="Defaults applied when creating a new commit."
-        >
-          <SettingsRow
-            id="commit.signoff"
-            label="Append Signed-off-by"
-            hint="Appends a DCO-style trailer to every commit message."
-            control={
-              <PGToggle
-                checked={s.addSignoff}
-                onChange={(v) => s.set("addSignoff", v)}
-              />
-            }
-          />
-          <SettingsRow
-            id="commit.ticket"
-            label="Ticket pattern"
-            hint={
-              <>
-                Regular expression run over the BRANCH NAME to find a ticket key
-                the commit composer offers as a one-click insert (#252). Capture
-                group 1 wins when the pattern has one, so{" "}
-                <code>issue-(\d+)</code> inserts just the number. Leave it empty
-                for no chip. Nothing is inserted automatically.
-              </>
-            }
-            control={
-              <PGInput
-                value={s.commitTicketPattern}
-                onChange={(v) => s.set("commitTicketPattern", v)}
-                // A pattern that will not compile means no chip and no
-                // explanation, so the field says so while it is being typed.
-                // `aria-invalid` alongside `error` because PGInput's `error` is
-                // a border colour and nothing more — adding the attribute to the
-                // shared primitive would restate the semantics of every input in
-                // the app in a change that is not about that.
-                error={!isValidTicketPattern(s.commitTicketPattern)}
-                aria-invalid={!isValidTicketPattern(s.commitTicketPattern)}
-                mono
-                size="sm"
-                placeholder={DEFAULT_TICKET_PATTERN}
-                style={{ width: 220 }}
-                data-testid="settings-ticket-pattern"
-              />
-            }
-          />
-          <SettingsRow
-            id="commit.sign"
-            label="Sign commits"
-            hint="Uses gpg.format, user.signingkey and gpg.program. Following git config respects commit.gpgsign per repository; a signing failure fails the commit rather than producing an unsigned one."
-            control={
-              <PGSelect
-                value={s.signCommits}
-                onChange={(v) =>
-                  s.set("signCommits", v as "config" | "always" | "never")
-                }
-                options={[
-                  { value: "config", label: "Follow git config" },
-                  { value: "always", label: "Always" },
-                  { value: "never", label: "Never" },
-                ]}
-              />
-            }
-          />
-        </SettingsCard>
+        <RemotePage />
 
         <DiffPage />
 
-        <ForgeSettings />
+        <IntegrationsPage />
         <KeyboardPage />
         <CliSection />
         <UpdatesPage />
@@ -324,46 +147,6 @@ export function SettingsScreen() {
         <DiagnosticsSection />
       </div>
     </div>
-  );
-}
-
-/**
- * `user.name` / `user.email` (#212) — the one thing on this screen that is NOT
- * a platypusgit preference.
- *
- * It is here anyway, and first, because it is the only setting the app cannot
- * work without: git refuses to record a commit until both are set, and until
- * #212 there was nowhere in the app to set them. The subtitle says out loud
- * that this writes git's own config, since the page's own header promises
- * "preferences are saved locally".
- *
- * Reachable with no repository open, which is why `repoId` is optional all the
- * way down: a user who lands in Settings before opening anything still gets a
- * true answer, from the global + system chain.
- */
-function IdentitySection() {
-  const repo = useRepoStore((s) => s.current);
-  return (
-    <SettingsCard
-      id="identity"
-      title="Identity"
-      subtitle="Who your commits are recorded as. Unlike everything else here, this is written to your git config — the same user.name and user.email git itself reads."
-    >
-      <SettingsRow
-        id="identity.author"
-        label="Commit author"
-        hint="git refuses to record a commit without both. The scope control decides whether saving writes this repository's own config or your global one."
-        stacked
-        control={<IdentityForm repoId={repo?.id ?? null} />}
-      />
-      <SettingsRow
-        id="identity.saved"
-        label="Saved identities"
-        hint="Keep the identities you switch between — a work address and a personal one. Applying one writes it to the OPEN repository's config, so git and every hook agree with what you see here. Editing or removing an entry does not change repositories that already use it."
-        stacked
-        control={<SavedIdentities repoId={repo?.id ?? null} />}
-      />
-    </SettingsCard>
   );
 }
